More explicit types

+++ lrf_serial_comm.c
@@ -170,7 +170,7 @@ static void on_uart_irq_callback(FuriHalSerialHandle *hndl,
     space is encountered */
 void strcpy_rstrip(char *dst, uint8_t *src) {
 
-  int i;
+  uint8_t i;
 
   for(i = 0; src[i] > 32 && src[i] < 127; i++)
     dst[i] = src[i];
@@ -183,7 +183,7 @@ void strcpy_rstrip(char *dst, uint8_t *src) {
 static uint8_t checkbyte(uint8_t *data, uint16_t len) {
 
   uint8_t checksum = 0;
-  int i;
+  uint16_t i;
 
   for(i = 0; i < len; i++)
     checksum += data[i];
